requirements.txt
python-magic
aiohttp[speedups]
